♻️[#278] TravelButtonView로 분리
